requirements.txt
Flask
python-dotenv
requests
openai
gspread
oauth2client
SQLAlchemy
rapidfuzz
